+++ UserImagesService/IUserService.cs
@@ -10,11 +10,11 @@ namespace UserImagesService
     public interface IUserService
     {
         IQueryable<User> GetUsers();
-        User GetUser(long id);
+        User GetUser(int id);
         IQueryable<User> FindByCondition(Expression<Func<User, bool>> expression);
         void InsertUser(User user);
         void UpdateUser(User user);
-        void DeleteUser(long id);
+        void DeleteUser(int id);
         
     }
 }
